+++ s/ok/fns/err.ts
@@ -2,7 +2,7 @@
 import type {Err} from "../types/err.js"
 
 /** failure */
-export function err<E = string>(error: E): Err<E> {
+export function err<E>(error: E): Err<E> {
 	return {ok: false, error}
 }