+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### v0.2.31
 - 🍏 add `attempt` fn for ok/err wrapping
+- 🍏 add `errorString` fn
 
 ### v0.2.30
 - 🍏 fix BaseX utils like hex/base58 to support leading zeros